A small refactoring

+++ Pod/Classes/UIColor+ColorWithHex.h
@@ -25,6 +25,6 @@
 
 @interface UIColor (ColorFromHex)
 
-+(UIColor*)colorWithHex:(int)hex;
++(UIColor*)colorWithHex:(unsigned long)hex;
 
 @end
